Rename public namespace to store
