+++ webapp/store/views.py
@@ -2,8 +2,8 @@ import flask
 import humanize
 import webapp.metrics.helper as metrics_helper
 import webapp.metrics.metrics as metrics
-from webapp.api.public import StoreApi
-import webapp.public.logic as logic
+from webapp.api.store import StoreApi
+import webapp.store.logic as logic
 from dateutil import parser
 from math import floor
 from webapp.api.exceptions import (
